Bug 1581079 - Try to fix crash in [@ PL_strncpy | PL_strncat | nsParseMailMessageState::GetAggregateHeader]. r=BenC
This clarifies the type of a few variables.
Ensure header value+length are not out of sync in a couple of places.

Differential Revision: https://phabricator.services.mozilla.com/D221032

+++ mailnews/local/src/nsParseMailbox.h
@@ -33,7 +33,7 @@ class nsIMsgFolder;
  */
 typedef struct message_header {
   const char* value; /* The contents of a header (after ": ") */
-  int32_t length;    /* The length of the data (it is not NULL-terminated.) */
+  size_t length;     /* The length of the data (it is not NULL-terminated.) */
 } message_header;
 
 // This object maintains the parse state for a single mail message.
